Registration form is enhanced and working (if issues related to e.g. concurrency, error logging and reporting etc. are excluded) but contains no JavaScript-based validation, Admin Site is configured, database model is enhanced and modified to follow user profile idiom, a new table for states (as in USA) is added and fixtures for countries, states and initial users are created.
A clean database can be created by deleting the old one and executing commands "syncdb" and "migrate albumizer" with the manage.py. When syncdb asks about creating superusers, the answer is no: the default users exist in the database fixtures. More users can be created afterwards as needed.

Data models are updated so that the albumizer's old user table is now UserProfile, and every albumizer's table referencing to a user has a direct relationship to the auth.User table - not the current UserProfile, as it used to be. A reference to user's UserProfile can be obtained in code by calling user.get_profile() method. The UserProfile tuple for each new user is created automagically by event handler tied to the auth.User table (see models.py). Also, models are enhanced with additional information like ordering, indexes, constraints, plain text names and help strings.