Started to implement a database populator as an additional command to manage.py. At the moment, it can be used to generate test data for everything else but content of albums. After the way of storing album content is cleared, the program can be extended to insert content to the albums it creates, too.
Help of the command can be obtained by issuing "python manage.py help insert_albumizer_test_data" (the command is also listed on the main help page). Amount of the data generated can be adjusted with a couple of parameters, as well as the verbosity of the command.

Generation of data can be started by executing e.g. "python manage.py insert_albumizer_test_data --verbosity=1 --users=5 --albumsmax=4".

After starting to use this command, everyone should switch to use local database, which is not stored to version control. With this command, huge amounts of test data can be generated, but the database files grow so big, that Git's pulls and pushes will become slow if the database files are included. A good way of doing the switch is to override the default database settings in developer-specific settings file, pointing the database file to be located somewhere else than under the project folder. After making the settings, a new sqlite db can be generated by syncdb/migrate, after which this data generation command can be utilized.

If a large test database is wanted, the command can be left to crunch the disk while doing something else... :-)

The code of the populator is not neat, but that's not so important in an utility program like the command is.