Refactored the test data populator to be more efficient when generating order information, so that it works quite smoothly when there are thousands of users in the database. Also modified handling of orders so, that all but the most recent orders are in state 3 (= ?) and the most recent one is in state 0 (= in shopping basket). Generation of album titles and descriptions is enhanced to include html tags to some content to make sure that nothing is going to crash when handling content including them.
Btw: The "official" database on the production server will be PostgreSQL, so it would be good to test the application with it. Performance of SQLite sucks, anyway, so I'd recommend to install Postgre. It works fine on my Win7, and output lines of the database populator like flash on the screen compared to how it was with SQLite. It is healthier for the hard disk, too. :-) Just install the Psycopg driver (http://initd.org/psycopg/) after Postgre, add an user for Albumizer, create a db owned by that user, and point Django to it in the personal settings file. After that the usual syncdb and migrate.