Some small improvements, essentially:
The Admin Site is now actually working with big amounts of data. Representation of some fields had to be changed to avoid loading of huge data masses.

Admin Site has settings for editing Facebook profiles.

The former method for returning json from view functions is converted into view decorator.

Added escaping of html tags to the json data about Albums.

Removed the flash-splash from the welcome page.

Modified the welcome page template and the related css so that the tabs and the twitter button on the page will become visible only after the page is loaded and the elements are rendered the right way.