Okay, the reason for the big memory consumption of the test data generator seems to be found: When executing in DEBUG mode, Django is saving a copy of each and every SQL statement it executes. That's why the app takes more and more memory without any visible reason, until the memory runs out. When generating thousands of persons with full related data, the number of db queries made would be quite big... :-P
Fortunately it took only one function call to fix that, so that the SQL statement record is cleared after each generated user. Now the memory consumption seems to spin around 20 Mbytes no matter how long the program is run. Of course the ids of generated user are still stored in a list, but that shouldn't be a problem until the amount of users is *huge*.