Shopping cart is essentially working and accessible albums containing pages can be added to it. Users' own albums can also be deleted. The underlying data model of shopping cart still needs to be refactored a little bit to separate it from the Album model to make it possible for albums to stay in the shopping cart after the album is deleted, until the cart is displayed next time.
There is a new show_messages tag, which lists messages delivered by Django's messages facility. It should be embedded on every page, preferably under the main heading but sometimes also above it. The functionality of this tag can be tested e.g. with the shopping cart or by deleting albums.