Selection of delivery addresses during the checkout process is implemented. One field is added to the ShoppingCartItem model, so please migrate your database.
Some server-side caching is added to the templates and the json api.

Some small bugs are fixed.

Some super-hyper-fancy jQuery effects added to some page elements either to make their appearance smoother or just to show that we know about the effects (something not to do in the real life...). :-)